# swagger.yaml
openapi: 3.0.0
info:
  title: Invoice API
  version: 2.0.0
  description: API for creating, managing, and tracking invoices. Make sure you select the vercel server link (Production server) so that you can actually use it.

servers:
  - url: https://mission-control-amber.vercel.app
    description: Production server

paths:

  /v1/auth/register:
    post:
      summary: Register a new user
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInput'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - missing email or password, or email already in use

  /v1/auth/login:
    post:
      summary: Login with email and password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthInput'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
        '400':
          description: Bad request - missing email or password
        '401':
          description: Unauthorised - invalid email or password

  /v1/auth/logout:
    post:
      summary: Logout current user
      tags: [Auth]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Logged out successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorised - no token provided

  /v1/invoices:
    post:
      summary: Create a draft invoice
      tags: [Invoices]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              type: string
      responses:
        '201':
          description: Invoice created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invoiceId:
                    type: string
                  filePath:
                    type: string
        '400':
          description: Bad request - missing required fields, invalid date formats, or invalid item values
        '401':
          description: Unauthorised - not logged in
        '403':
          description: Forbidden - no permission to create invoices

  /v1/invoices/{id}:
    get:
      summary: Retrieve an invoice by ID
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invoice'
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found - invoice ID does not exist

    put:
      summary: Update a draft invoice
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvoiceInput'
      responses:
        '200':
          description: Invoice updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - invalid values
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found
        '409':
          description: Conflict - invoice already finalised

    delete:
      summary: Delete a draft invoice
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found - invoice does not exist
        '409':
          description: Conflict - invoice already finalised, sent, or paid

  /v1/invoices/{id}/status:
    get:
      summary: Get invoice status and lifecycle tracking
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Invoice status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvoiceWithStatus'
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found - invoice ID does not exist
        '409':
          description: Conflict - invalid lifecycle state detected

    put:
      summary: Update invoice status
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [Generated, InProgress, Sent, Paid, Overdue, Deleted]
      responses:
        '200':
          description: Status updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '400':
          description: Bad request - invalid status value
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found - invoice ID does not exist

  /v1/invoices/{id}/pdf:
    get:
      summary: Download invoice as PDF
      tags: [Invoices]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: PDF file returned successfully
          content:
            application/pdf:
              schema:
                type: string
                format: binary
        '401':
          description: Unauthorised
        '403':
          description: Forbidden
        '404':
          description: Not found - invoice or PDF not found
        '409':
          description: Conflict - invoice is not finalised yet, PDF not available

  /v1/ai/chat:
    post:
      summary: Send a message to the AI invoicing assistant
      tags: [AI]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message:
                  type: string
      responses:
        '200':
          description: AI response returned successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tier:
                    type: string
                    enum: [standard, pro]
                  messagesRemaining:
                    type: integer
                    nullable: true
        '400':
          description: Bad request - missing message
        '401':
          description: Unauthorised
        '404':
          description: User not found
        '429':
          description: Daily message limit reached - upgrade to Pro for unlimited messages

  /v1/ai/chat/history:
    delete:
      summary: Clear AI chat history (Pro users only)
      tags: [AI]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Chat history cleared successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorised
        '403':
          description: Forbidden - Pro tier required

  /v1/ai/invoice/generate:
    post:
      summary: Generate an invoice from a natural language description (Standard and Pro)
      tags: [AI]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: Create an invoice for Sarah Johnson for 10 hours of cloud consulting at $250 per hour
      responses:
        '201':
          description: Invoice generated and saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invoiceId:
                    type: string
                  invoiceData:
                    $ref: '#/components/schemas/InvoiceInput'
        '400':
          description: Bad request - missing description
        '401':
          description: Unauthorised
        '404':
          description: User not found
        '429':
          description: Daily message limit reached - upgrade to Pro for unlimited messages
        '500':
          description: AI failed to generate valid invoice data

  /v1/ai/invoice/autofill:
    post:
      summary: Generate invoice field suggestions from a natural language description (Pro only)
      tags: [AI]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: Bill Sarah Johnson at Horizon Digital for 10 hours of consulting at $250 per hour
      responses:
        '200':
          description: Invoice fields generated successfully for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invoiceData:
                    $ref: '#/components/schemas/InvoiceInput'
        '400':
          description: Bad request - missing description
        '401':
          description: Unauthorised
        '403':
          description: Forbidden - Pro tier required
        '404':
          description: User not found
        '500':
          description: AI failed to generate valid invoice data

  /v1/users/tier:
    put:
      summary: Update user tier
      tags: [Users]
      parameters:
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tier]
              properties:
                tier:
                  type: string
                  enum: [standard, pro]
      responses:
        '200':
          description: Tier updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  tier:
                    type: string
        '400':
          description: Bad request - invalid tier or missing tier
        '401':
          description: Unauthorised
        '404':
          description: User not found

  /v1/ai/invoice/update/{id}:
    post:
      summary: Update an existing invoice using AI (Pro only)
      tags: [AI]
      parameters:
        - $ref: '#/components/parameters/InvoiceId'
        - in: header
          name: token
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [description]
              properties:
                description:
                  type: string
                  example: Change the customer name to Jane Smith and update the rate to $300 per hour
      responses:
        '200':
          description: Updated invoice fields returned for review
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  invoiceData:
                    $ref: '#/components/schemas/InvoiceInput'
        '400':
          description: Bad request - missing description
        '401':
          description: Unauthorised
        '403':
          description: Forbidden - Pro tier required or invoice belongs to another user
        '404':
          description: Not found - invoice does not exist
        '409':
          description: Conflict - invoice is already finalised
        '500':
          description: AI failed to generate valid invoice data

components:

  parameters:
    InvoiceId:
      name: id
      in: path
      required: true
      schema:
        type: string
      description: The unique invoice ID

  schemas:

    AuthInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    Address:
      type: object
      required: [street, city, postcode, country]
      properties:
        street:
          type: string
        city:
          type: string
        postcode:
          type: string
        country:
          type: string

    CustomerInformation:
      type: object
      required: [fullName, email, phone, billingAddress, shippingAddress]
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        billingAddress:
          $ref: '#/components/schemas/Address'
        shippingAddress:
          $ref: '#/components/schemas/Address'

    FromDetails:
      type: object
      required: [businessName, address, taxId, abnNumber]
      properties:
        businessName:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        taxId:
          type: string
        abnNumber:
          type: string

    LineItem:
      type: object
      required: [description, quantity, rate]
      properties:
        description:
          type: string
        quantity:
          type: number
        rate:
          type: number

    TaxDetails:
      type: object
      required: [taxId, countryCode, taxPercentage]
      properties:
        taxId:
          type: string
        countryCode:
          type: string
        taxPercentage:
          type: number

    InvoiceInput:
      type: object
      properties:
        customer:
          $ref: '#/components/schemas/CustomerInformation'
        lineItems:
          type: array
          items:
            $ref: '#/components/schemas/LineItem'
        currency:
          type: string
          default: AUD
        exchangeRate:
          type: number
        baseCurrency:
          type: string
        tax:
          $ref: '#/components/schemas/TaxDetails'
        from:
          $ref: '#/components/schemas/FromDetails'

    Invoice:
      type: object
      properties:
        invoiceId:
          type: string
        userId:
          type: string
        xml:
          type: string
        status:
          type: string
          enum: [Generated, InProgress, Sent, Paid, Overdue, Deleted]

    InvoiceWithStatus:
      type: object
      properties:
        invoiceId:
          type: string
        userId:
          type: string
        xml:
          type: string
        status:
          type: string
          enum: [Generated, InProgress, Sent, Paid, Overdue, Deleted]
        createdAt:
          type: string
          format: date-time